Created a specrum match parent plot
